pyproject.toml
[project]
name = "paper-scraper"
version = "0.1.0"
description = "ExamCooker paper validator with OCR and a review dashboard."
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "fastapi>=0.116.0",
  "httpx>=0.28.1",
  "jinja2>=3.1.6",
  "pillow>=11.3.0",
  "pymupdf>=1.26.4",
  "pytesseract>=0.3.13",
  "rapidfuzz>=3.14.0",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.2",
]

[project.scripts]
paper-audit = "paper_audit.cli:main"

[build-system]
requires = ["setuptools>=80.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["paper_audit*"]
